--- a/levenshtein/csharp/code.cs
+++ b/levenshtein/csharp/code.cs
@@ -18,32 +18,43 @@ for (int i = 0; i < args.Length; i++)
 Console.WriteLine($"times: {times}");
 Console.WriteLine($"min_distance: {min_distance}");
 
-static int levenshtein(string str1, string str2)
+static int levenshtein(ReadOnlySpan<char> str1t, ReadOnlySpan<char> str2t)
 {
-    int n = str1.Length;
-    int m = str2.Length;
-    int[,] matrix = new int[m + 1, n + 1];
+    // Get lengths of both strings
+  int mt = str1t.Length;
+  int nt = str2t.Length;
+  // Assign shorter one to str1, longer one to str2
+  ReadOnlySpan<char> str1 = mt <= nt ? str1t : str2t;
+  ReadOnlySpan<char> str2 = mt <= nt ? str2t : str1t;
+  // store the lengths of shorter in m, longer in n
+  int m = str1 == str1t ? mt : nt;
+  int n = str1 == str1t ? nt : mt;
+ 
+  // Create two rows, previous and current
+  Span<int> prev = stackalloc int[m+1];
+  Span<int> curr = stackalloc int[m+1];
+ 
+  // initialize the previous row
+  for (int i = 0; i <= m; i++) {
+    prev[i] = i;
+  }
 
-    for (int i = 0; i <= m; i++)
-    {
-        matrix[i, 0] = i;
+  // Iterate and compute distance
+  for (int i = 1; i <= n; i++) {
+    curr[0] = i;
+    for (int j = 1; j <= m; j++) {
+      int cost = (str1[j-1] == str2[i-1]) ? 0 : 1;
+      curr[j] = Math.Min(
+        prev[j] + 1,      // Deletion
+        Math.Min(curr[j-1] + 1,    // Insertion
+        prev[j-1] + cost)  // Substitution
+      );
     }
-
-    for (int j = 0; j <= n; j++)
-    {
-        matrix[0, j] = j;
+    for (int j = 0; j <= m; j++) {
+      prev[j] = curr[j];
     }
-
-    for (int i = 1; i <= m; i++)
-    {
-        for (int j = 1; j <= n; j++)
-        {
-            int cost = (str1[i - 1] == str2[j - 1] ? 0 : 1);
-            matrix[i, j] = int.Min(
-                matrix[i - 1, j] + 1, int.Min(matrix[i, j - 1] + 1,
-                          matrix[i - 1, j - 1] + cost));
-        }
-    }
-
-    return matrix[m, n];
+  }
+  
+  // Return final distance, stored in prev[m]
+  return prev[m];
 }