1- Replaced `string` with `ReadOnlySpan<char>` in Levenshtein function to avoid additional allocations and improve efficiency for slicing strings.
2- Use stackalloc instead of new.