@Ichoran 2024-12-22
This looks good to me--much more like the reference implementation!