@Akeit0 2024-12-26
You can add ``[MethodImpl(MethodImplOptions.AggressiveOptimization)]`` to levenshtein,
And you can use swapping instead of coping.
```cs
 // Swap spans
var temp = prev;
prev = curr;
curr = temp;
```