@PEZ 2024-12-30
@AhmedZero I'm sorry for the radio silence. We're doing this as a side project in addition to working for food and taking care of our families. 😀

I'll try to give your changes a look tomorrow. 

I'm on my phone and can't check right now, but if you haven't attached benchmark results from before and after your changes, please do that. It makes it much easier for us to assess things.