@AhmedZero 2024-12-30
Old version:
Output:
```
Checking C#
Check passed
Benchmarking C#
Benchmark 1: ./csharp/code/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa ...
  Time (mean ± σ):      82.9 ms ±   1.3 ms    [User: 72.9 ms, System: 7.6 ms]
  Range (min … max):    81.5 ms …  83.9 ms    3 runs
```

New Version:
```
Checking C#
Check passed
Benchmarking C#
Benchmark 1: ./csharp/code/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa ...
  Time (mean ± σ):      52.0 ms ±   1.7 ms    [User: 40.1 ms, System: 10.5 ms]
  Range (min … max):    50.3 ms …  53.8 ms    3 runs
```

new version is approximately 37.27% faster than the old version.