@PEZ 2024-12-30
🎉 On my machine this also fixes so that I can actually run the executable. The previous code resulted in some index out of bounds and then just hanged.

Also about that performance gain. It is probably much more than 37%. Since we have the CLR start time in the mix. I'm now very curious about the AOT build. I fail to compile it on my machine:

```
dotnet publish csharp -o csharp/code-aot /p:PublishAot=true /p:OptimizationPreference=Speed
Restore complete (0.5s)
  code failed with 1 error(s) (4.7s) → csharp/bin/Release/net9.0/osx-arm64/code.dll
    /Users/pez/.nuget/packages/microsoft.dotnet.ilcompiler/9.0.0/build/Microsoft.NETCore.Native.targets(376,5): error MSB3073: The command ""clang" -ld_classic "obj/Release/net9.0/osx-arm64/native/code.o" -o "bin/Release/net9.0/osx-arm64/native/code" -exported_symbols_list "obj/Release/net9.0/osx-arm64/native/code.exports" -Wl,-dead_strip -gz=zlib /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/sdk/libbootstrapper.o /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/sdk/libRuntime.WorkstationGC.a /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/sdk/libeventpipe-disabled.a /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/sdk/libstandalonegc-disabled.a /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/sdk/libstdc++compat.a /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/framework/libSystem.Native.a /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/framework/libSystem.Globalization.Native.a /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/framework/libSystem.IO.Compression.Native.a /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/framework/libSystem.Net.Security.Native.a /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/framework/libSystem.Security.Cryptography.Native.Apple.a /opt/homebrew/Cellar/dotnet/9.0.101/libexec/packs/runtime.osx-arm64.Microsoft.DotNet.ILCompiler/9.0.0/framework/libSystem.Security.Cryptography.Native.OpenSsl.a --target=arm64-apple-macos12.0 -g -Wl,-rpath,'@executable_path' -lssl -lcrypto -lbrotlienc -lbrotlidec -ldl -lobjc -lswiftCore -lswiftFoundation -lz -licucore -lm -L/usr/lib/swift -framework CoreFoundation -framework CryptoKit -framework Foundation -framework Security -framework GSS" exited with code 1.

Build failed with 1 error(s) in 5.3s
```

If anyone knows what I can do to fix this, that'd be awesome!