@AhmedZero 2024-12-30
Is clang installed?